:bug: Fix delete

+++ lndb_setup/_delete.py
@@ -10,34 +10,25 @@ from ._settings_store import instance_settings_file
 
 def delete(instance_name: str):
     """Delete an instance."""
+    logger.info(f"Deleting instance {settings.user.handle}/{instance_name}")
     settings_file = instance_settings_file(instance_name, settings.user.handle)
     isettings = load_instance_settings(settings_file)
 
-    if isettings.is_remote:
-        logger.info("Please delete your remote instance on lamin.ai.")
-    else:
-        if isettings.storage.type == "local":
-            delete_storage(isettings.storage.root)
-        else:
-            delete_cache(isettings.cache_dir)
-            logger.info("Storage won't be deleted as it is a cloud storage.")
-
     delete_settings(settings_file)
-
-
-def delete_storage(storage_root: Path):
-    if storage_root.exists():
-        shutil.rmtree(storage_root)
-        logger.info("Instance default storage root deleted.")
+    delete_cache(isettings.cache_dir)
+    logger.info(
+        f"    consider deleting your stored data manually: {isettings.storage.root}"
+    )
+    if isettings.is_remote:
+        logger.info("    please manually delete your remote instance on lamin.ai")
 
 
 def delete_cache(cache_dir: Path):
-    if cache_dir:
-        if cache_dir.exists():
-            shutil.rmtree(cache_dir)
-            logger.info("Instance cache deleted.")
+    if cache_dir is not None and cache_dir.exists():
+        shutil.rmtree(cache_dir)
+        logger.info("    instance cache deleted")
 
 
 def delete_settings(settings_file: Path):
     settings_file.unlink()
-    logger.info("Instance settings deleted.")
+    logger.info("    instance settings deleted")
